requirements.txt
boto3==1.36.26
botocore==1.36.26
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
google-api-core==2.24.2
google-api-python-client==2.163.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.69.1
greenlet==3.1.1
gspread==6.2.0
httplib2==0.22.0
idna==3.10
jmespath==1.0.1
numpy==2.2.3
oauth2client==4.1.3
oauthlib==3.2.2
pandas==2.2.3
proto-plus==1.26.1
protobuf==5.29.3
psycopg2-binary==2.9.10
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
s3transfer==0.11.2
six==1.17.0
SQLAlchemy==1.4.36
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
